+++ src/test/scala/edu/rit/cs/mmior/jsonoid/discovery/JsonSchemaSpec.scala
@@ -125,7 +125,7 @@ class JsonSchemaSpec extends UnitSpec {
       JsonSchema.fromJson(objSchema).asInstanceOf[ObjectSchema]
 
     convertedSchema.properties should contain(
-      StaticDependenciesProperty(Map("foo" -> List("bar")))
+      StaticDependenciesProperty(Map("foo" -> Set("bar")))
     )
   }