+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Allow disabling expansion entirely
 - Add possible debug output to CLI
+- Add some compatibility checking for `StaticDependenciesProperty`
 
 ### Fixed
 - Avoid crash during expansion if a property is missing